+++ src/index.js
@@ -109,7 +109,6 @@ var createFromScriptTag = require('./lib/create-from-script-tag');
 var constants = require('./constants');
 var analytics = require('./lib/analytics');
 var DropinError = require('./lib/dropin-error');
-var Promise = require('./lib/promise');
 var wrapPromise = require('@braintree/wrap-promise');
 
 var VERSION = '__VERSION__';