+++ src/views/base-view.js
@@ -4,7 +4,6 @@ var assign = require('../lib/assign').assign;
 var classList = require('@braintree/class-list');
 var DropinError = require('../lib/dropin-error');
 var errors = require('../constants').errors;
-var Promise = require('../lib/promise');
 
 function BaseView(options) {
   options = options || {};